sdk/python-sdk/requirements.txt
wheel==0.33.4
setuptools==41.0.1
python3-indy==1.8.2
requests==2.22.0
pylint==2.3.1
pylint-quotes==0.2.1
pytest==3.6.4
pytest-parallel==0.0.9
